Promo modules

+++ Controllers/promotionController.js
@@ -25,36 +25,40 @@ class PromotionController {
     }
 
     async addPromotion(req, res, next) {
-        console.log("Données reçues:", req.body); // 👀 Vérifier ce qui est envoyé
 
-        const { title, training_id, students, trainers, supervisors } = req.body;
+        const { title, training_id, students, trainers, supervisors, modules, start_date, end_date } = req.body;
         try {
             if (!title || !training_id) {
                 throw new CustomError("Titre et formation obligatoires.", 400);
             }
+            //vérifier si le nom de la formation existe déjà
+            const existingPromotion = await promotionService.getPromotionByTitle(title);
+            if (existingPromotion.length > 0) {
+                throw new CustomError("Une promotion avec le même nom existe déjà.", 400);
+            }
 
-            const promotionData = { title, training_id };
+            const promotionData = { title, training_id, start_date, end_date};
 
-            const newPromotion = await promotionService.addPromotion(promotionData, students, trainers, supervisors);
+            const newPromotion = await promotionService.addPromotion(promotionData, students, trainers, supervisors, modules);
             res.status(201).json(newPromotion);
         } catch (error) {
-            next(new CustomError("Une erreur est survenue lors de l'ajout de la promotion.", 500));
+            next(error);
         }
     }
 
     async updatePromotion(req, res, next) {
         const { promotionId } = req.params;
-        const { title, training_id } = req.body;
+        const { title, start_date, end_date} = req.body;
         try {
             // Validation des données d'entrée
-            if (!title && !training_id) {
+            if (!title) {
                 throw new CustomError("Titre et formation obligatoires.", 400);
             }
 
-            const promotionData = { title, training_id };
+            const promotionData = { title, start_date, end_date};
 
             await promotionService.updatePromotion(promotionId, promotionData);
-            res.status(200).json({ message: "Promotion mise à jour." });
+            res.status(200).json({ message: `Promotion ${title} mise à jour.` });
         } catch (error) {
             next(error)
         }
@@ -63,10 +67,22 @@ class PromotionController {
     async deletePromotion(req, res, next) {
         const { promotionId } = req.params;
         try {
+            const existingUserPromotion = await promotionService.findPromotion(promotionId);
+
+            if (
+                existingUserPromotion.studients.length > 0 ||
+                existingUserPromotion.trainers.length > 0 ||
+                existingUserPromotion.supervisors.length > 0
+            ) {
+                throw new CustomError("Cette promotion contient un ou plusieurs utilisateurs.", 400);
+            }
+
+
             await promotionService.deletePromotion(promotionId);
             res.status(200).json({ message: "Promotion supprimée." });
+
         } catch (error) {
-            next(error)
+            next(error);
         }
     }
